+++ src/types/interfaces/services/core/context/types.ts
@@ -7,9 +7,7 @@
  * We're starting with simple base types for the core context.
  * The base types will expand as we determine what is common across all services.
  */
-// eslint-disable-next-line @typescript-eslint/no-empty-object-type
 export type CoreCommand<TCommand extends object = object> = TCommand;
 // eslint-disable-next-line @typescript-eslint/no-empty-object-type
 export type CoreState<TState extends {} = {}> = TState;
-// eslint-disable-next-line @typescript-eslint/no-empty-object-type
 export type CoreUpdate<TUpdate = unknown> = TUpdate;